# docs/swagger.yaml
definitions:
  models.AuthBookRequest:
    properties:
      book_id:
        type: integer
    type: object
  models.AuthStatus:
    properties:
      display_name:
        type: string
      link_token:
        type: string
      link_token_ttl_sec:
        type: integer
      status:
        type: string
      student_group:
        type: string
    type: object
  models.AuthTokens:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
...
  models.CreateEventRequest:
    properties:
      description:
        type: string
      event_date:
        type: string
      event_type_code:
        type: integer
      icon_url:
        type: string
      points:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.CreateEventResponse:
    properties:
      created:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_id:
        type: integer
      event_type_code:
        type: integer
      icon_url:
        type: string
      points:
        type: integer
      title:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.Event:
    properties:
      created:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_id:
        type: integer
      event_type_code:
        type: integer
      icon_url:
        type: string
      points:
        type: integer
      title:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      auth:
        $ref: '#/definitions/models.AuthTokens'
      user:
        $ref: '#/definitions/models.UserSubstructure'
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - token
    type: object
  models.RegisterResponse:
    properties:
      auth:
        $ref: '#/definitions/models.AuthTokens'
      user:
        $ref: '#/definitions/models.UserSubstructure'
    type: object
  models.ResetPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.SetNewPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
      successful:
        type: boolean
    type: object
...
  models.UserSubstructure:
    properties:
      book_id:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      role_level:
        type: integer
      student_group:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /admin/create_event:
    post:
      consumes:
      - application/json
      description: Создаёт новое событие в системе и возвращает полную информацию
        о созданной записи. Требует авторизации администратора. Поля помимо title
        могут не указываться, тогда они будут заменены на стандартные значения
      parameters:
      - description: 'Bearer токен авторизации. Формат: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные нового события
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Событие успешно создано
          schema:
            $ref: '#/definitions/models.CreateEventResponse'
        "400":
          description: Некорректный JSON или ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Нет доступа — невалидный или отсутствующий токен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при создании события
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать событие
      tags:
      - admin
...
  /admin/events:
    get:
      consumes:
      - application/json
      description: Возвращает полный список событий из базы данных.
      parameters:
      - description: 'Bearer токен авторизации. Формат: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список событий
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "500":
          description: Ошибка сервера при получении списка событий
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить все события
      tags:
      - admin
...
  /auth/check:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет наличие студенческого билета в системе.
        Если студент не зарегистрирован, генерирует временный токен для регистрации.
      parameters:
      - description: Номер студенческого билета
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AuthBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Студенческий найден, выдан регистрационный токен
          schema:
            $ref: '#/definitions/models.AuthStatus'
        "400":
          description: Некорректный запрос — ошибка парсинга JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Студенческий не найден в базе
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Пользователь с таким номером уже зарегистрирован
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при генерации или сохранении токена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Проверка студенческого в системе
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет почту и пароль пользователя. При успешной авторизации
        выдает пару access и refresh токенов.
      parameters:
      - description: Данные для входа (почта и пароль)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация — возвращается пользователь и пара токенов
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Некорректный запрос — ошибка парсинга JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неправильный пароль
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь с такой почтой не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при работе с базой данных или токенами
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя на основе временного токена, выданного
        после проверки студенческого билета.
      parameters:
      - description: Почта, пароль и регистрационный токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Некорректный формат JSON или токен истёк
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Студент с таким номером студенческого не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Пользователь с такой почтой уже существует
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера (база данных, хеширование, транзакция)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация пользователя по токену
      tags:
      - auth
  /auth/reset_password:
    post:
      consumes:
      - application/json
      description: |-
        Отправляет на почту пользователя ссылку для сброса пароля.
        Если пользователь с указанной почтой существует — ему придёт письмо с временной ссылкой на установку нового пароля.
      parameters:
      - description: Почта пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Инструкция отправлена на почту
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при поиске пользователя или отправке письма
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Запрос на сброс пароля
      tags:
      - auth
  /auth/set_new_password:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по токену из письма. Токен действителен
        ограниченное время (15 минут).
      parameters:
      - description: Токен и новый пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SetNewPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно обновлён
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Некорректный JSON
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Невалидный или истекший токен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при обновлении пароля
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Установка нового пароля
      tags:
      - auth
...
swagger: "2.0"